Re-import failed messages from error.log on a fresh db
If for some reason the user has deleted the RavenDB database and starting fresh and the user attempts to move messages from error.log back into the error queue for ServiceControl to rehydrate the error messages, in this case, we won't have a corresponding saga. Therefore we are using SingleOrDefault instead of Single.